# docs/swagger.yaml
basePath: /v1
definitions:
...
  dto.TemplateCreateRequest:
    properties:
      description:
        type: string
      exercises:
        items:
          $ref: '#/definitions/dto.TemplateExerciseItem'
        type: array
      name:
        maxLength: 100
        type: string
    required:
    - name
    type: object
  dto.TemplateExerciseItem:
    properties:
      exercise_id:
        type: string
      planned_sets:
        type: integer
    required:
    - exercise_id
    type: object
  dto.TemplateExerciseResponse:
    properties:
      exercise_id:
        type: string
      planned_sets:
        type: integer
      sort_order:
        type: integer
    type: object
  dto.TemplateMediaResponse:
    properties:
      id:
        type: string
      media_type:
        type: string
      s3_key:
        type: string
      sort_order:
        type: integer
    type: object
  dto.TemplateResponse:
    properties:
      created_at:
        type: string
      created_by_user_id:
        type: string
      description:
        type: string
      exercises:
        items:
          $ref: '#/definitions/dto.TemplateExerciseResponse'
        type: array
      id:
        type: string
      is_public:
        type: boolean
      media:
        items:
          $ref: '#/definitions/dto.TemplateMediaResponse'
        type: array
      name:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
  dto.TemplateUpdateRequest:
    properties:
      description:
        type: string
      exercises:
        items:
          $ref: '#/definitions/dto.TemplateExerciseItem'
        type: array
      name:
        maxLength: 100
        type: string
    type: object
...
  handler.ErrorDetail:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/handler.ValidationDetail'
        type: array
      message:
        type: string
      user_message:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/handler.ErrorDetail'
    type: object
  handler.ValidationDetail:
    properties:
      field:
        type: string
      param:
        type: string
      tag:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: Core gym domain service for OneRep — exercises, templates, workouts,
    progress
  title: OneRep Gym API
  version: "1.0"
paths:
...
  /templates:
    get:
      consumes:
      - application/json
      description: List the authenticated user's templates, optionally filtered by
        last-updated time
      parameters:
      - description: Only templates updated after this RFC 3339 timestamp
        in: query
        name: since
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TemplateResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List templates
      tags:
      - templates
    post:
      consumes:
      - application/json
      description: Create a custom template with a list of exercises and planned sets
      parameters:
      - description: Template data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TemplateCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TemplateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a template
      tags:
      - templates
  /templates/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-delete an owned template
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a template
      tags:
      - templates
    get:
      consumes:
      - application/json
      description: Get a template by ID, including its exercises and media
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TemplateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a template
      tags:
      - templates
    patch:
      consumes:
      - application/json
      description: Patch an existing template; null fields are left unchanged
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TemplateUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TemplateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a template
      tags:
      - templates
  /templates/{id}/fork:
    post:
      consumes:
      - application/json
      description: Copy a public template, or an own private template, including its
        exercises and media, as the authenticated user's own
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TemplateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Fork a template
      tags:
      - templates
  /templates/{id}/publish:
    post:
      consumes:
      - application/json
      description: Make an owned template publicly visible to other users
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TemplateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Publish a template
      tags:
      - templates
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"